requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyserial==3.0.1
Werkzeug==0.11.3
wheel==0.24.0
